Move KubernetesModelITs
