+++ server/src/integration/java/org/opentosca/toscana/plugins/kubernetes/model/KubernetesExpressIT.java
@@ -1,4 +1,4 @@
-package org.opentosca.toscana.plugins.kubernetes;
+package org.opentosca.toscana.plugins.kubernetes.model;
 
 import java.io.File;
